init DB to allow connection to postgres on pg from client on barman

+++ scorch.sh
@@ -1,2 +1,7 @@
 #!/bin/sh
-sudo rm -rf var_lib_postgresql var_log
+set -eu
+if [ -z "$(sudo find var_lib_postgresql -name *.pid -print)" ]; then
+  echo scorching var_lib_postgresql var_log
+  sudo rm -rf var_lib_postgresql var_log
+fi
+
